Add FileHandler mixin and models module.
