+++ iscc_web/api/models.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+from pydantic import BaseModel
+
+
+class UploadMeta(BaseModel):
+
+    file_name: str
+    content_type: str
+    client_ip: str
